Don't duplicate Chrome MV2 code in MV3
