+++ shared/js/ui/base/chrome-ui-wrapper.es6.js
@@ -1,22 +1,24 @@
-let fetch = (message) => {
+const parseUserAgentString = require('../../shared-utils/parse-user-agent-string.es6.js')
+
+const fetch = (message) => {
     return new Promise((resolve, reject) => {
-        window.chrome.runtime.sendMessage(message, (result) => resolve(result))
+        chrome.runtime.sendMessage(message, (result) => resolve(result))
     })
 }
 
-let backgroundMessage = (thisModel) => {
+const backgroundMessage = (thisModel) => {
     // listen for messages from background and
     // // notify subscribers
-    window.chrome.runtime.onMessage.addListener((req, sender) => {
+    chrome.runtime.onMessage.addListener((req, sender) => {
         if (sender.id !== chrome.runtime.id) return
         if (req.whitelistChanged) thisModel.send('whitelistChanged')
         if (req.updateTabData) thisModel.send('updateTabData')
         if (req.didResetTrackersData) thisModel.send('didResetTrackersData', req.didResetTrackersData)
-        if (req.closePopup) window.close()
+        if (req.closePopup) closePopup()
     })
 }
 
-let getBackgroundTabData = () => {
+const getBackgroundTabData = () => {
     return new Promise((resolve, reject) => {
         fetch({getCurrentTab: true}).then((tab) => {
             if (tab) {
@@ -28,34 +30,34 @@ let getBackgroundTabData = () => {
     })
 }
 
-let search = (url) => {
-    window.chrome.tabs.create({url: `https://duckduckgo.com/?q=${url}&bext=${window.localStorage['os']}cr`})
+const search = (url) => {
+    const os = parseUserAgentString().os
+    chrome.tabs.create({url: `https://duckduckgo.com/?q=${url}&bext=${os}cr`})
 }
 
-let getExtensionURL = (path) => {
+const getExtensionURL = (path) => {
     return chrome.extension.getURL(path)
 }
 
-let openExtensionPage = (path) => {
-    window.chrome.tabs.create({ url: getExtensionURL(path) })
+const openExtensionPage = (path) => {
+    chrome.tabs.create({ url: getExtensionURL(path) })
 }
 
-let openOptionsPage = (browser) => {
+const openOptionsPage = (browser) => {
     if (browser === 'moz') {
         openExtensionPage('/html/options.html')
-        window.close()
+        closePopup()
     } else if (browser === 'chrome') {
-        window.chrome.runtime.openOptionsPage()
+        chrome.runtime.openOptionsPage()
     }
 }
 
-let reloadTab = (id) => {
-    window.chrome.tabs.reload(id)
+const reloadTab = (id) => {
+    chrome.tabs.reload(id)
 }
 
-let closePopup = () => {
-    const w = window.chrome.extension.getViews({type: 'popup'})[0]
-    w.close()
+const closePopup = () => {
+    window.close()
 }
 
 module.exports = {
